+++ CHANGELOG.md
@@ -27,8 +27,8 @@
   * Nest and Unnest have been renamed to Wrap and Unwrap respectively. This is
     to better conform to TUTORIAL D's terminology.
     
-  * Lispy#chain was kept public in 0.9.0 and has been tagged as private as not
-    being part of the DSL. 
+  * Lispy#chain was kept public in 0.9.0 by error and has been entirely removed 
+    from the DSL.
 
 * Enhancements (internals)